improve script to check errors

+++ scripts/install.sh
@@ -1,7 +1,21 @@
+geotemporal_has() {
+  type "$1" > /dev/null 2>&1
+}
+
 SO=$(uname -s)
 DUCKDB_VERSION=0.4.0
 
 mkdir -p /var/lib/libduckdb
+
+if ! geotemporal_has "curl"; then
+    echo "Error: curl command not found"
+    exit 1
+fi
+if [ "$EUID" -ne 0 ]
+  then echo "Please run as root to install libraries in /var/lib/"
+  exit
+fi
+
 
 if [ "$SO" == "Darwin" ]; then
     rm -rf /usr/local/bin/geotemporal-data-explorer
@@ -14,8 +28,13 @@ else
     curl https://storage.googleapis.com/geotemporal-data-explorer-releases/versions/geotemporal-data-explorer-linux-amd64-latest?rand=$RANDOM --output /bin/geotemporal-data-explorer
     chmod +x /bin/geotemporal-data-explorer
 fi
-unzip -u /var/lib/libduckdb/libduckdb.zip -d /var/lib/libduckdb
-rm -rf /var/lib/libduckdb/libduckdb.zip
+if geotemporal_has "unzip"; then
+    unzip -u /var/lib/libduckdb/libduckdb.zip -d /var/lib/libduckdb 
+    rm -rf /var/lib/libduckdb/libduckdb.zip
+else
+    echo "Error: unzip command not found"
+    exit 1
+fi
 
 echo "Close and reopen your terminal to start using geotemporal-data-explorer or load your
 preferences again"
